refactor: move vehicle pages into designated folders
