+++ src/pages-components/Dashboard/FleetPage/Trailers/SingleTrailerPage/GeneralInfo.tsx
