+++ src/pages-components/Dashboard/FleetPage/Trailers/SingleTrailerPage/InfoItem.tsx
