+++ src/pages-components/Dashboard/FleetPage/Trailers/SingleTrailerPage/index.ts
