+++ src/pages-components/Dashboard/FleetPage/Trailers/TrailersPage/TrailersPage.tsx
@@ -1,38 +1,41 @@
 import { DashboardLayout } from '@/components/layout/DashboardLayout';
 import type { Vehicle } from '@/lib/api';
 import { LoadingPage } from '@/lib/components/LoadingPage';
-import { useTrucks } from '@/lib/hooks';
-import { TrucksTable } from '@/pages-components/Dashboard/FleetPage/TrucksPage/TrucksTable';
+import { useTrailers } from '@/lib/hooks';
 import { Box, Button, Divider, FlexLayout, Heading, Text } from '@/ui';
 import pluralize from 'pluralize';
 
-export const TrucksPage = () => {
-  const { trucks, isLoading } = useTrucks();
+import { TrailersTable } from './TrailersTable';
 
-  return <DashboardLayout>{isLoading || !trucks ? <LoadingPage /> : <MainContent trucks={trucks} />}</DashboardLayout>;
+export const TrailersPage = () => {
+  const { trailers, isLoading } = useTrailers();
+
+  return (
+    <DashboardLayout>{isLoading || !trailers ? <LoadingPage /> : <MainContent trailers={trailers} />}</DashboardLayout>
+  );
 };
 
-const MainContent = ({ trucks }: { trucks: Vehicle[] }) => {
+const MainContent = ({ trailers }: { trailers: Vehicle[] }) => {
   return (
     <Box>
       <Box className="flex items-center justify-between">
         <FlexLayout className="flex-col gap-2">
           <Heading as="h1" variant="text-xl">
-            Fleet — Trucks
+            Fleet — Trailers
           </Heading>
           <Divider />
           <Text color="text-color-2" variant="text-m">
             There are{' '}
             <Text variant="text-m-bold">
-              {trucks.length} {pluralize('Truck', trucks.length)}
+              {trailers.length} {pluralize('Trailers', trailers.length)}
             </Text>{' '}
             in your fleet
           </Text>
         </FlexLayout>
-        <Button isDisabled href="/dashboard/fleet/trucks/new" iconLeft="PlusIcon" text="New Truck" />
+        <Button isDisabled href="/dashboard/fleet/trailers/new" iconLeft="PlusIcon" text="New Trailer" />
       </Box>
       <Box className="py-5">
-        <TrucksTable trucks={trucks} />
+        <TrailersTable trailers={trailers} />
       </Box>
     </Box>
   );
