+++ src/pages-components/Dashboard/FleetPage/Trucks/TrucksPage/TrucksPage.tsx
@@ -1,40 +1,39 @@
 import { DashboardLayout } from '@/components/layout/DashboardLayout';
 import type { Vehicle } from '@/lib/api';
 import { LoadingPage } from '@/lib/components/LoadingPage';
-import { useTrailers } from '@/lib/hooks';
-import { TrailersTable } from '@/pages-components/Dashboard/FleetPage/TrailersPage/TrailersTable';
+import { useTrucks } from '@/lib/hooks';
 import { Box, Button, Divider, FlexLayout, Heading, Text } from '@/ui';
 import pluralize from 'pluralize';
 
-export const TrailersPage = () => {
-  const { trailers, isLoading } = useTrailers();
+import { TrucksTable } from './TrucksTable';
 
-  return (
-    <DashboardLayout>{isLoading || !trailers ? <LoadingPage /> : <MainContent trailers={trailers} />}</DashboardLayout>
-  );
+export const TrucksPage = () => {
+  const { trucks, isLoading } = useTrucks();
+
+  return <DashboardLayout>{isLoading || !trucks ? <LoadingPage /> : <MainContent trucks={trucks} />}</DashboardLayout>;
 };
 
-const MainContent = ({ trailers }: { trailers: Vehicle[] }) => {
+const MainContent = ({ trucks }: { trucks: Vehicle[] }) => {
   return (
     <Box>
       <Box className="flex items-center justify-between">
         <FlexLayout className="flex-col gap-2">
           <Heading as="h1" variant="text-xl">
-            Fleet — Trailers
+            Fleet — Trucks
           </Heading>
           <Divider />
           <Text color="text-color-2" variant="text-m">
             There are{' '}
             <Text variant="text-m-bold">
-              {trailers.length} {pluralize('Trailers', trailers.length)}
+              {trucks.length} {pluralize('Truck', trucks.length)}
             </Text>{' '}
             in your fleet
           </Text>
         </FlexLayout>
-        <Button isDisabled href="/dashboard/fleet/trailers/new" iconLeft="PlusIcon" text="New Trailer" />
+        <Button isDisabled href="/dashboard/fleet/trucks/new" iconLeft="PlusIcon" text="New Truck" />
       </Box>
       <Box className="py-5">
-        <TrailersTable trailers={trailers} />
+        <TrucksTable trucks={trucks} />
       </Box>
     </Box>
   );
