+++ src/pages/dashboard/fleet/trailers/index.ts
@@ -1,1 +1,1 @@
-export { TrailersPage as default } from '@/pages-components/Dashboard/FleetPage/TrailersPage';
+export { TrailersPage as default } from '@/pages-components/Dashboard/FleetPage/Trailers/TrailersPage';
