+++ src/pages/dashboard/fleet/trucks/[id]/index.ts
@@ -1,1 +1,1 @@
-export { SingleTruckPage as default } from '@/pages-components/Dashboard/FleetPage/SingleTruckPage';
+export { SingleTruckPage as default } from '@/pages-components/Dashboard/FleetPage/Trucks/SingleTruckPage';
