+++ src/pages/dashboard/fleet/trucks/index.ts
@@ -1,1 +1,1 @@
-export { TrucksPage as default } from '@/pages-components/Dashboard/FleetPage/TrucksPage';
+export { TrucksPage as default } from '@/pages-components/Dashboard/FleetPage/Trucks/TrucksPage';
